Upload score on death

+++ Objects/Enemies/Computer/Entity.cs
@@ -181,7 +181,7 @@ namespace Debugmancer.Objects.Enemies.Computer
 			Modulate = new Color(1, 1, 1);
 			if (health == 0)
 			{
-				Globals.Score += Math.Ceiling(100 * Globals.ScoreMultiplier);
+				Globals.Score += (int)Math.Ceiling(100 * Globals.ScoreMultiplier);
 				GetParent().GetNode<KinematicBody2D>("Player").GetNode<Label>("HUD/Score").Text = $"Score: {Globals.Score}";
 				QueueFree();
 			}
