+++ Objects/Enemies/Void/Entity.cs
@@ -111,7 +111,7 @@ namespace Debugmancer.Objects.Enemies.Void
 			Modulate = new Color(1, 1, 1);
 			if (health == 0)
 			{
-				Globals.Score += Math.Ceiling(125 * Globals.ScoreMultiplier);
+				Globals.Score += (int)Math.Ceiling(125 * Globals.ScoreMultiplier);
 				GetParent().GetNode<KinematicBody2D>("Player").GetNode<Label>("HUD/Score").Text = $"Score: {Globals.Score}";
 				ChangeState("Dead");
 			}
