+++ Objects/Player/Globals.cs
@@ -2,7 +2,7 @@ namespace Debugmancer.Objects.Player
 {
     public class Globals
     {
-        public static double Score = 0;
+        public static int Score = 0;
         public static double ScoreMultiplier = 1;
         public static bool CanDash = false;
         public static int PlayerDamage = 1;
